fix: change id type to int64 from int

+++ get_movie_external_ids.go
@@ -6,14 +6,14 @@ import (
 )
 
 type MovieExternalIDs struct {
-	ID          int     `json:"id"`
+	ID          int64   `json:"id"`
 	ImdbID      *string `json:"imdb_id"`
 	FacebookID  *string `json:"facebook_id"`
 	InstagramID *string `json:"instagram_id"`
 	TwitterID   *string `json:"twitter_id"`
 }
 
-func (client *Client) MovieExternalIDs(ctx context.Context, movieID int) (*MovieExternalIDs, error) {
+func (client *Client) MovieExternalIDs(ctx context.Context, movieID int64) (*MovieExternalIDs, error) {
 	url := fmt.Sprintf("%s/movie/%d/external_ids?api_key=%s", client.baseURL, movieID, client.apiKey)
 	external := new(MovieExternalIDs)
 	if err := client.get(ctx, url, external); err != nil {
